ServiceBadge->badge_hash 改成 bigint

+++ webdata/models/ServiceBadge.php
@@ -11,7 +11,7 @@ class ServiceBadge extends Pix_Table
         $this->_columns['service_id'] = ['type' => 'int'];
         $this->_columns['service_user'] = ['type' => 'int'];
         $this->_columns['badge_time'] = ['type' => 'int'];
-        $this->_columns['badge_hash'] = ['type' => 'int'];
+        $this->_columns['badge_hash'] = ['type' => 'bigint'];
         $this->_columns['brief'] = ['type' => 'text'];
         $this->_columns['data'] = ['type' => 'jsonb'];
